refactor(labels): split the encoder into buffer, string table and encoder
The encoder was one class holding three unrelated things: a growable
byte buffer, an open-addressed string interning table, and the knowledge
of what jfr_labels.proto looks like. Eight mutable fields, nineteen
methods, and no way to test or replace any one part.

Now ProtoBuffer owns the bytes and the varint primitives, StringTable
owns interning and id assignment, and LabelsSnapshotEncoder owns only
the schema: field numbers, nesting, record sizes. Behaviour is
unchanged, including the constant-seeding semantics.

The split had to not cost anything, because the hot path's whole trick
is one capacity check per record and then raw stores into hoisted
locals. So ProtoBuffer exposes reserve/array/commit rather than a method
per byte, and callers keep writing into a local `byte[] b` with a local
position. Measured on the two stress-test shapes, 2 forks x 20
iterations each:

  160k contexts x 2 labels    34.654 -> 33.193 ms/op   (+-2.2, +-1.9)
  10240 contexts x 20 labels  27.959 -> 26.718 ms/op   (+-1.7, +-1.8)
  allocation, both shapes     20894558 -> 20894570 B/op

Both timings landed nominally faster with heavily overlapping error
bars, so read it as no measurable change. The +12 B/op is the two extra
object headers, against 21 MB per dump.

Also drops a stringId() delegate that nothing called.
